CyberSentinel/requirements.txt
flask
python-dotenv
google-genai
requests
pefile
reportlab